+++ app/public/javascripts/collections.js
@@ -1,0 +1,10 @@
+
+var DayCollection = Backbone.Collection.extend({
+    model: DayModel,
+    url: '/october'
+});
+
+var BookingCollection = Backbone.Collection.extend({
+  model: BookingModel,
+  url: '/booking'
+})
