+++ app/views/calendar.hbs
@@ -4,7 +4,7 @@
 	<meta charset="utf-8">
 	<title>User Dashboard</title>
 	 <meta name="viewport" content="width=device-width, initial scale=1.0">
-	<link rel="stylesheet" href="../public/stylesheets/calendarStyle.css" media="screen">
+	<link rel="stylesheet" href="/stylesheets/calendarStyle.css" media="screen">
 </head>
 <body>
 
@@ -69,7 +69,7 @@
 </div>
 
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js"></script>
-<script src="calendar.js"></script>
+<script src="/javascripts/calendar.js"></script>
 	<script>
 	var calendar = Calendar();
 	calendar.init();
