+++ app/views/main.hbs
@@ -11,10 +11,10 @@
   </form>
 </body>
 </html> -->
-<script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js"></script>
+<!-- <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js"></script>
 <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.8.3/underscore-min.js"></script>
-<script type='text/javascript' src='https://cdnjs.cloudflare.com/ajax/libs/backbone.js/1.2.3/backbone-min.js'></script>
-<script type="text/javascript" src="javascripts/main.js"></script>
+<script type='text/javascript' src='https://cdnjs.cloudflare.com/ajax/libs/backbone.js/1.2.3/backbone-min.js'></script> -->
+
 <!--this script is needed to get rid of the '#_=_' hash that Facebook sends with the callback-->
 <script type="text/javascript">
     if (window.location.hash && window.location.hash == '#_=_') {
@@ -47,6 +47,7 @@
       <p>First Name: {{user.firstName}}</p>
       <p>Last Name: {{user.lastName}}</p>
     </div>
+    <form id= "seeCalendar" action = "/calendar" method="post"><input type="submit" value = "See Calendar"></form>
   </div>
 
 {{else}}
@@ -60,3 +61,4 @@
   </div>
 
 {{/if}}
+<script type="text/javascript" src="javascripts/main.js"></script>
